apps/api/requirements/base.txt
Django>=5.0,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
channels>=4.0
channels-redis>=4.2
daphne>=4.1
social-auth-app-django>=5.4
social-auth-core>=4.5
psycopg2-binary>=2.9
resend>=2.0
upstash-redis>=1.0
python-json-logger>=2.0
sentry-sdk[django]>=1.40
Pillow>=10.0
requests>=2.31
